added venue rating, mock venue creation

+++ src/main/java/fipu/diplomski/dmaglica/repo/entity/UserEntity.kt
@@ -7,12 +7,17 @@ import jakarta.persistence.*
 class UserEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    var id: Long = 0
+    var id: Int = 0
 
     var username: String = ""
+
     var email: String = ""
+
     var password: String = ""
+
     var lastKnownLatitude: Double? = null
+
     var lastKnownLongitude: Double? = null
+
     var roleId: Int = 0
 }
